Finished up while and added quotient for test cases and fixed a bug in cond
cond wouldn't work in this case: (cond [else 8]) which was actually one
of the tests surprisingly enough.
Also: while now works and when we expand named let and letrec it will no
longer be part of eval-exp. I only mention this because it is the only
thing stopping eval-exp from being cps